+++ src/settings/GobiIntegration/GobiIntegration.test.js
@@ -1,5 +1,9 @@
-import { render, screen } from '@testing-library/react';
 import { MemoryRouter } from 'react-router-dom';
+
+import {
+  render,
+  screen,
+} from '@folio/jest-config-stripes/testing-library/react';
 
 import GobiIntegration from './GobiIntegration';