+++ src/settings/components/FieldFromFieldName/FieldFromFieldName.test.js
@@ -1,6 +1,11 @@
-import user from '@testing-library/user-event';
-import { act, render, screen } from '@testing-library/react';
 import { Form } from 'react-final-form';
+
+import {
+  act,
+  render,
+  screen,
+} from '@folio/jest-config-stripes/testing-library/react';
+import userEvent from '@folio/jest-config-stripes/testing-library/user-event';
 
 import { getFieldNameOptions } from '../../MappingConfiguration/utils';
 import { FieldFromFieldName } from './FieldFromFieldName';
@@ -34,10 +39,8 @@ describe('FieldFromFieldName', () => {
   it('should change \'Path\' when gobi field name was changed', async () => {
     renderFieldFromFieldName();
 
-    await act(async () => {
-      await user.click(screen.getByText('stripes-components.selection.controlLabel'));
-      await user.click(screen.getByText(defaultProps.dataOptions[1].label));
-    });
+    await act(async () => userEvent.click(screen.getByText('stripes-components.selection.controlLabel')));
+    await act(async () => userEvent.click(screen.getByText(defaultProps.dataOptions[1].label)));
 
     expect(defaultProps.change).toHaveBeenCalled();
   });