+++ src/settings/constants/mapping.js
@@ -91,6 +91,7 @@ export const FIELDS = {
   SUBSCRIPTION_INTERVAL: 'SUBSCRIPTION_INTERVAL',
   SUBSCRIPTION_TO: 'SUBSCRIPTION_TO',
   SUFFIX: 'SUFFIX',
+  SUPPRESS_INSTANCE_FROM_DISCOVERY: 'SUPPRESS_INSTANCE_FROM_DISCOVERY',
   TAGS: 'TAGS',
   TITLE: 'TITLE',
   TOTAL_ESTIMATED_PRICE: 'TOTAL_ESTIMATED_PRICE',
@@ -129,6 +130,7 @@ export const GOBI_FIELDS = {
   localData2: 'Local Data 2',
   localData3: 'Local Data 3',
   localData4: 'Local Data 4',
+  suppressInstanceFromDiscovery: 'Suppress instance from discovery',
 };
 
 export const GOBI_FIELDS_PATH_MAP = {
@@ -151,4 +153,5 @@ export const GOBI_FIELDS_PATH_MAP = {
   [GOBI_FIELDS.localData2]: '//LocalData[Description=\'LocalData2\']/Value',
   [GOBI_FIELDS.localData3]: '//LocalData[Description=\'LocalData3\']/Value',
   [GOBI_FIELDS.localData4]: '//LocalData[Description=\'LocalData4\']/Value',
+  [GOBI_FIELDS.suppressInstanceFromDiscovery]: '//SuppressInstanceFromDiscovery',
 };